Added topup function to admin extension
