+++ lnbits/core/services.py
@@ -122,16 +122,4 @@ def check_invoice_status(wallet_id: str, payment_hash: str) -> PaymentStatus:
     return WALLET.get_invoice_status(payment.checking_id)
 
 
-def update_wallet_balance(wallet_id: str, amount: int) -> str:
-    temp_id = f"temp_{urlsafe_short_hash()}"
-    internal_id = f"internal_{urlsafe_short_hash()}"
-    create_payment(
-        wallet_id=wallet_id,
-        checking_id=internal_id,
-        payment_request="admin_internal",
-        payment_hash="admin_internal",
-        amount=amount * 1000,
-        memo="Admin top up",
-        pending=False,
-    )
-    return "admin_internal"
+
