+++ lnbits/extensions/admin/crud.py
@@ -1,0 +1,17 @@
+from lnbits.helpers import urlsafe_short_hash
+from lnbits.core.crud import create_payment
+
+
+def update_wallet_balance(wallet_id: str, amount: int) -> str:
+    temp_id = f"temp_{urlsafe_short_hash()}"
+    internal_id = f"internal_{urlsafe_short_hash()}"
+    create_payment(
+        wallet_id=wallet_id,
+        checking_id=internal_id,
+        payment_request="admin_internal",
+        payment_hash="admin_internal",
+        amount=amount * 1000,
+        memo="Admin top up",
+        pending=False,
+    )
+    return "success"
